+++ SharePointFramework/ProjectExtensions/src/projectSetup/index.ts
@@ -29,6 +29,7 @@ import { IProjectSetupData, IProjectSetupProperties, ProjectSetupValidation } fr
 import { find } from 'underscore'
 import { endsWith } from 'lodash'
 import { ProjectSetupSettings } from './ProjectSetupSettings'
+import { ProjectTemplateFile } from 'models/ProjectTemplateFile'
 
 export default class ProjectSetup extends BaseApplicationCustomizer<IProjectSetupProperties> {
   private _portal: PortalDataService
@@ -402,7 +403,6 @@ export default class ProjectSetup extends BaseApplicationCustomizer<IProjectSetu
         </Query>
     </View>`
       }
-
       const [templates, extensions, listContentConfig] = await Promise.all([
         this._portal.getItems(
           this.properties.templatesLibrary,
@@ -410,7 +410,7 @@ export default class ProjectSetup extends BaseApplicationCustomizer<IProjectSetu
           {
             ViewXml: templateViewXml
           },
-          ['File', 'FieldValuesAsText']
+          ['File','FieldValuesAsText']
         ),
         this.properties.extensionsLibrary
           ? this._portal.getItems(
@@ -424,9 +424,14 @@ export default class ProjectSetup extends BaseApplicationCustomizer<IProjectSetu
             )
           : Promise.resolve([]),
         this.properties.contentConfigList
-          ? this._portal.getItems(this.properties.contentConfigList, ListContentConfig)
+          ? this._portal.getItems(this.properties.contentConfigList, ListContentConfig, {}, ['File'])
           : Promise.resolve([]),
       ])
+      const files = await this._portal.getItems(strings.Lists_ProjectTemplateFiles_Title, ProjectTemplateFile, {}, ['File'])
+       templates.map((template) => {
+         const [relativeUrl] = files.filter(file => {return file.id === template.projectTemplateId})
+         template.serverRelativeUrl = relativeUrl.serverRelativeUrl
+       })
       Logger.log({
         message: '(ProjectSetup) [_fetchData]: Retrieved templates, extensions and content config',
         data: {